// OTA-KU/frontend/src/api/generated/models/ListTerminateForOTA.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ListTerminateForOTA = {
  /**
   * ID mahasiswa asuh
   */
  mahasiswaId: string;
  maName: string;
  maNIM: string;
  requestTerminationNoteOTA: string;
  requestTerminationNoteMA: string;
  requestTerminateMa: boolean;
  requestTerminateOta: boolean;
  createdAt: string;
};